Disabled persisting connection with Redis Cache.
1. Converted Typeorm db connection to async/await
2. Restructured redis cache files.
