+++ pmt-auth/package.json
@@ -5,8 +5,8 @@
   "main": "app.js",
   "scripts": {
     "build": "tsc --project './tsconfig.json'",
-    "dev": "tsc --project './tsconfig.json' --watch & NODE_ENV=development nodemon build/app.js",
-    "prod": "tsc --project './tsconfig.json' --watch & NODE_ENV=development node build/app.js"
+    "start:dev": "tsc --project './tsconfig.json' --watch & NODE_ENV=development nodemon build/app.js",
+    "start:prod": "tsc --project './tsconfig.json' --watch & NODE_ENV=development node build/app.js"
   },
   "author": "DevSoc",
   "license": "MIT",
