+++ pmt-auth/src/app.ts
@@ -38,7 +38,6 @@ class App {
 		} else {
 			app.loadServer();
 			app.loadDatabase();
-			app.loadCache();
 
 			console.log(`Worker ${process.pid} started`);
 		}
